Move master and music bots into seperate files
